feat:add cache for department and schedule service
