+++ src/main/java/com/asclepius/AsclepiusApplication.java
@@ -3,11 +3,13 @@ package com.asclepius;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cache.annotation.EnableCaching;
 
 /**
  * @author luolinyuan
  */
 @SpringBootApplication
+@EnableCaching
 @MapperScan("com.asclepius.mapper")
 public class AsclepiusApplication {
