+++ src/main/java/com/asclepius/dto/AppointmentDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.io.Serializable;
+
 /**
  * @Author sny
  * @CreateTime 2022-07-01  19:28
@@ -13,7 +15,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class AppointmentDTO {
+public class AppointmentDTO implements Serializable {
 
     private Integer apId;
