+++ src/main/java/com/asclepius/dto/DoctorDTO.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.io.Serializable;
+
 /**
  * @Author sny
  * @CreateTime 2022-07-01  15:54
@@ -13,7 +15,7 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class DoctorDTO {
+public class DoctorDTO implements Serializable {
     private Integer docId;
 
     private String docName;
