+++ src/main/resources/application-deploy.properties
@@ -8,6 +8,8 @@ mybatis.configuration.map-underscore-to-camel-case=true
 mybatis.type-aliases-package=com.asclepius.pojo
 mybatis.mapper-locations=classpath:mapper/*.xml
 
+spring.cache.type=redis
+
 #redis\u4E3B\u673A\u5730\u5740(\u5982\u679C\u4F60\u5728\u672C\u5730\u542F\u52A8\u5C31\u662F127.0.0.1\uFF0C\u5982\u679C\u5728\u865A\u62DF\u673A\u5C31\u586B\u4F60\u865A\u62DF\u673A\u7684\u5730\u5740)
 spring.redis.host=49.232.196.77
 #redis\u7AEF\u53E3\uFF08\u9ED8\u8BA46379\uFF09
