+++ src/main/resources/application-dev.properties
@@ -1,6 +1,7 @@
 server.port=8080
 spring.application.name=asclepius
 spring.datasource.url=jdbc:mysql://localhost:3306/asclepius
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.datasource.username=root
 spring.datasource.password=qwe123
 
@@ -8,6 +9,8 @@ mybatis.configuration.map-underscore-to-camel-case=true
 mybatis.type-aliases-package=com.asclepius.pojo
 mybatis.mapper-locations=classpath:mapper/*.xml
 mybatis.configuration.log-impl=org.apache.ibatis.logging.log4j.Log4jImpl
+
+spring.cache.type=redis
 
 #redis\u4E3B\u673A\u5730\u5740(\u5982\u679C\u4F60\u5728\u672C\u5730\u542F\u52A8\u5C31\u662F127.0.0.1\uFF0C\u5982\u679C\u5728\u865A\u62DF\u673A\u5C31\u586B\u4F60\u865A\u62DF\u673A\u7684\u5730\u5740)
 spring.redis.host=localhost
