+++ 接口说明文档.md
@@ -90,7 +90,7 @@
 
 ```json
 {
-	"code": 422,
+	"code": 403,
     "message": "",
     "data": null
 }
@@ -98,14 +98,13 @@
 
 ### 1.2.2 验证token是否有效
 
-- 已废弃
-- 请求路径：login/validity/{token}
+- 请求路径：login/validity
 - 请求方法：get
 - 请求参数
 
-| 参数名 | 参数说明 | 备注     |
-| ------ | -------- | -------- |
-| token  | token    | 不可为空 |
+| 参数名 | 参数说明                    | 备注     |
+| ------ | --------------------------- | -------- |
+| token  | 放入请求头的AUTHORIZATION中 | 不可为空 |
 
 - 有效
 
@@ -122,7 +121,7 @@
 ```json
 {
 	"code": 403,
-   "message": "yes",
+   "message": "",
    "data": null
 }
 ```
@@ -221,11 +220,15 @@
 ### 1.3.3 删除就诊卡
 
 - [ ] 前端
-- [ ] 后端
-- [ ] 测试
-  - 请求路径：user/deleteCard/{uId}/{cId}
 
+- [ ] 后端
+
+- [ ] 测试
+  
+  请求路径：user/deleteCard/{uId}/{cId}
+  
 - 请求方法：delete
+
 - 请求参数：
 
 | 参数名 | 参数说明 | 备注     |
@@ -262,7 +265,7 @@
 }
 ```
 
-### 1.3.4 修改就诊卡信息
+### 1.3.4 修改就诊卡信息(已废弃)
 
 - [ ] 前端
 - [ ] 后端
